Modularize RSA blind signature logic
